src/chatbot/requirements.txt
openai
discord.py
boto3
pytz
geopy
google-api-python-client
beautifulsoup4
sumy
tinysegmenter
spacy
ginza
ja-ginza
tiktoken
